+++ src/main/java/christmas/domain/event/BenefitCalculator.java
@@ -1,7 +1,5 @@
-package christmas.service;
+package christmas.domain.event;
 
-import christmas.domain.event.EventDetail;
-import christmas.domain.event.MatchingEvents;
 import christmas.domain.orders.Orders;
 import christmas.domain.visitingDate.VisitingDate;
